Move components to components directory and store project images in public folder
